+++ packages/image_bytes_cache/test/unit_test.dart
@@ -9,6 +9,7 @@ import 'open/open_test.dart' as open_test;
 import 'resolve/bearer_middleware_test.dart' as bearer_middleware_test;
 import 'resolve/client_test.dart' as client_test;
 import 'resolve/conditional_middleware_test.dart' as conditional_middleware_test;
+import 'resolve/freshness_test.dart' as freshness_test;
 import 'resolve/logger_middleware_test.dart' as logger_middleware_test;
 import 'resolve/pipeline_test.dart' as pipeline_test;
 import 'resolve/resolver_test.dart' as resolver_test;
@@ -27,6 +28,7 @@ void main() => group('Unit', () {
   indexed_cache_test.main();
   open_test.main();
   resolver_test.main();
+  freshness_test.main();
   client_test.main();
   pipeline_test.main();
   bearer_middleware_test.main();